ABC-213: focus message box after switching channel (#726)
The changes in PLT-8478 subtly changed the `id` of the AutosizeTextarea
from `post_textbox` to `post_textbox-textarea`. Unfortunately, the
former is hard-coded in a variety of places that trigger the desired
focus behaviour.

Ideally, this would be abstracted (as commented in the code), but for
now the minor portion of this change is reverted with no observable side
effects.