ABC-244: fix team invite link handling
The `hash` and `data` parameters, being `null`, were cast to strings
downstream in mattermost-redux as part of `buildQueryString`, resulting
in a value of `null` being passed up instead of the expected empty
string. The server correctly rejected the hash as invalid.

This change copies the more correct handling from the login controller.
Perhaps an even better solution would be to update `buildQueryString` in
mattermost-redux to check `null` parameters and send up empty strings
automatically, however that's probably not the change we want near the
end of a release. I'll propose the change in `master` separately
instead.