Remove mm_config and mm_license global state from webapp (PR #2) (#819)
* MM-8589: expunge global mm_(config|license) from integration components

This also reorganizes `components/integrations/components` as just
`components/integrations`

* MM-8589: expunge global mm_config from post_view components

* MM-8589: expunge global mm_config from admin_console components

* fix missing key on admin_console RadioSetting component

* MM-8589: expunge global mm_license from admin_console components

This commit takes a different tack than others on account of the heavy
use of the license object throughout the pages and the non-connectedness
of most components. Instead of denormalizing the individual properties
of the license per components, the `license` object itself is passed
down in most cases. Unfortunately, this means the `=== 'true'` checks
remain littered throughout, but those will be removed at the same time
we cut over all the connectors.

* MM-8589: expunge unnecessary ownProps

* MM-8589: use the get(Config|License) selector