MM-9973: ignore dispatched events against unregistered suggestion boxes (#1052)
In a slow network environment, typing a channel name, then erasing it,
then quickly typing again and hitting enter to unmount the suggestion
box could result in a previous network call eventually dispatching an
event for the now unregistered suggestion box. If the dispatched event
was for an empty string, it would incorrectly try to handle that,
ultimately attempting to dereference `undefined.push`.