MM-16152: always focus RHS on SELECT_POST (#2987)
Ideally, we'd refactor focus to be an entity we can directly manipulate in a reducer, but that's a non-trivial refactoring. For now, I'm capturing the timestamp of the last `SELECT_POST` and using that within the component to decide when to ignore/effect a focus of the RHS textbox, given that we already track the blur timestamp.