GH-11568 Adds cypress test for inserting emoji via keyboard input (#3081)
* adds spec for adding emoji via keyboard input

* updates to id's and jquery selector

* removes randomness, updates selection patterns

This removes the randomness of selection and uses input to assert emoji selection in a square shape.  The first "right" is lost to grab of selection so that is why left is right-1.

Several class selectors were updated to ID's. This includes a couple new ID's defined within the emoji picker preview pane itself.

* separates test into a more complete suite

I decided to separate the tests out into more steps but still retain the full multi-block movement test that the issue requested.  This is the test that is prefixed by _M16738_.

* adds emojiPickerCategories id

* adds helper for testing against index of selected emoji